move `build_agg_call_from_prost` to `AggCall::from_protobuf`
