+++ src/expr/src/function/aggregate/call.rs
@@ -12,12 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::sync::Arc;
+
+use risingwave_common::bail;
 use risingwave_common::types::DataType;
-use risingwave_common::util::sort_util::ColumnOrder;
+use risingwave_common::util::sort_util::{ColumnOrder, OrderType};
+use risingwave_pb::expr::PbAggCall;
 
 use super::AggKind;
-use crate::expr::ExpressionRef;
+use crate::expr::{build_from_prost, ExpressionRef};
 use crate::function::args::FuncArgs;
+use crate::Result;
 
 /// Represents an aggregation function.
 #[derive(Clone, Debug)]
@@ -42,4 +47,46 @@ pub struct AggCall {
 
     /// Should deduplicate the input before aggregation.
     pub distinct: bool,
+}
+
+impl AggCall {
+    pub fn from_protobuf(agg_call: &PbAggCall, append_only: bool) -> Result<Self> {
+        let agg_kind = AggKind::from_protobuf(agg_call.get_type()?)?;
+        let args = match &agg_call.get_args()[..] {
+            [] => FuncArgs::None,
+            [arg] if agg_kind != AggKind::StringAgg => {
+                FuncArgs::Unary(DataType::from(arg.get_type()?), arg.get_index() as usize)
+            }
+            [agg_arg, extra_arg] if agg_kind == AggKind::StringAgg => FuncArgs::Binary(
+                [
+                    DataType::from(agg_arg.get_type()?),
+                    DataType::from(extra_arg.get_type()?),
+                ],
+                [agg_arg.get_index() as usize, extra_arg.get_index() as usize],
+            ),
+            _ => bail!("Too many/few arguments for {:?}", agg_kind),
+        };
+        let column_orders = agg_call
+            .get_order_by()
+            .iter()
+            .map(|col_order| {
+                let col_idx = col_order.get_column_index() as usize;
+                let order_type = OrderType::from_protobuf(col_order.get_order_type().unwrap());
+                ColumnOrder::new(col_idx, order_type)
+            })
+            .collect();
+        let filter = match agg_call.filter {
+            Some(ref pb_filter) => Some(Arc::from(build_from_prost(pb_filter)?)),
+            None => None,
+        };
+        Ok(AggCall {
+            kind: agg_kind,
+            args,
+            return_type: DataType::from(agg_call.get_return_type()?),
+            column_orders,
+            append_only,
+            filter,
+            distinct: agg_call.distinct,
+        })
+    }
 }
