+++ src/stream/src/from_proto/agg_common.rs
@@ -13,65 +13,14 @@
 // limitations under the License.
 
 use std::collections::HashMap;
-use std::convert::TryFrom;
 use std::sync::Arc;
 
-use risingwave_common::bail;
 use risingwave_common::buffer::Bitmap;
-use risingwave_common::types::DataType;
-use risingwave_common::util::sort_util::{ColumnOrder, OrderType};
-use risingwave_expr::expr::build_from_prost;
-use risingwave_expr::function::aggregate::{AggCall, AggKind};
-use risingwave_expr::function::args::FuncArgs;
-use risingwave_pb::expr::PbAggCall;
 
 use super::*;
 use crate::common::table::state_table::StateTable;
 use crate::common::StateTableColumnMapping;
 use crate::executor::aggregation::AggStateStorage;
-
-pub fn build_agg_call_from_prost(
-    append_only: bool,
-    agg_call_proto: &PbAggCall,
-) -> StreamResult<AggCall> {
-    let agg_kind = AggKind::from_protobuf(agg_call_proto.get_type()?)?;
-    let args = match &agg_call_proto.get_args()[..] {
-        [] => FuncArgs::None,
-        [arg] if agg_kind != AggKind::StringAgg => {
-            FuncArgs::Unary(DataType::from(arg.get_type()?), arg.get_index() as usize)
-        }
-        [agg_arg, extra_arg] if agg_kind == AggKind::StringAgg => FuncArgs::Binary(
-            [
-                DataType::from(agg_arg.get_type()?),
-                DataType::from(extra_arg.get_type()?),
-            ],
-            [agg_arg.get_index() as usize, extra_arg.get_index() as usize],
-        ),
-        _ => bail!("Too many/few arguments for {:?}", agg_kind),
-    };
-    let column_orders = agg_call_proto
-        .get_order_by()
-        .iter()
-        .map(|col_order| {
-            let col_idx = col_order.get_column_index() as usize;
-            let order_type = OrderType::from_protobuf(col_order.get_order_type().unwrap());
-            ColumnOrder::new(col_idx, order_type)
-        })
-        .collect();
-    let filter = match agg_call_proto.filter {
-        Some(ref prost_filter) => Some(Arc::from(build_from_prost(prost_filter)?)),
-        None => None,
-    };
-    Ok(AggCall {
-        kind: agg_kind,
-        args,
-        return_type: DataType::from(agg_call_proto.get_return_type()?),
-        column_orders,
-        append_only,
-        filter,
-        distinct: agg_call_proto.distinct,
-    })
-}
 
 /// Parse from stream proto plan agg call states, generate state tables and column mappings.
 /// The `vnodes` is generally `Some` for Hash Agg and `None` for Simple Agg.
