+++ src/stream/src/from_proto/local_simple_agg.rs
@@ -15,7 +15,6 @@
 use risingwave_expr::function::aggregate::AggCall;
 use risingwave_pb::stream_plan::SimpleAggNode;
 
-use super::agg_common::build_agg_call_from_prost;
 use super::*;
 use crate::executor::LocalSimpleAggExecutor;
 
@@ -35,7 +34,7 @@ impl ExecutorBuilder for LocalSimpleAggExecutorBuilder {
         let agg_calls: Vec<AggCall> = node
             .get_agg_calls()
             .iter()
-            .map(|agg_call| build_agg_call_from_prost(node.is_append_only, agg_call))
+            .map(|agg_call| AggCall::from_protobuf(agg_call, node.is_append_only))
             .try_collect()?;
 
         Ok(LocalSimpleAggExecutor::new(
